refactored matrix area into multiple components
