+++ libs/error-analysis/src/lib/ErrorAnalysisDashboard/Controls/MatrixArea/MatrixArea.styles.ts
@@ -8,19 +8,11 @@ import {
 } from "office-ui-fabric-react";
 
 export interface IMatrixAreaStyles {
+  emptyLabelPadding: IStyle;
   matrixLabelBottom: IStyle;
   matrixLabelTab: IStyle;
-  emptyLabelPadding: IStyle;
   matrixLabel: IStyle;
-  matrixCell: IStyle;
-  matrixCellPivot1Categories: IStyle;
-  matrixCellPivot2Categories: IStyle;
-  matrixCol: IStyle;
-  matrixRow: IStyle;
   matrixArea: IStyle;
-  nanMatrixCell: IStyle;
-  styledMatrixCell: IStyle;
-  selectedMatrixCell: IStyle;
 }
 
 export const matrixAreaStyles: () => IProcessedStyleSet<
@@ -35,51 +27,6 @@ export const matrixAreaStyles: () => IProcessedStyleSet<
       flexDirection: "row",
       paddingBottom: "50px",
       paddingTop: "10px"
-    },
-    matrixCell: {
-      ":focus": {
-        outline: "none"
-      },
-      alignItems: "center",
-      border: "1px solid #bfbfbf",
-      cursor: "pointer",
-      display: "flex",
-      fontSize: "10px",
-      margin: "2px",
-      minWidth: "60px",
-      overflow: "visible"
-    },
-    matrixCellPivot1Categories: {
-      alignItems: "center",
-      border: "none",
-      display: "flex",
-      fontSize: "12px",
-      justifyContent: "flex-end",
-      lineHeight: "16px",
-      margin: "2px",
-      overflow: "hidden",
-      paddingLeft: "2px",
-      paddingRight: "10px",
-      width: "180px"
-    },
-    matrixCellPivot2Categories: {
-      alignItems: "center",
-      border: "1px transparent solid",
-      display: "flex",
-      fontSize: "12px",
-      justifyContent: "center",
-      lineHeight: "16px",
-      margin: "2px",
-      maxWidth: "60px",
-      minWidth: "60px",
-      overflow: "hidden"
-    },
-    matrixCol: {
-      display: "flex",
-      flexDirection: "column",
-      fontFamily: "Segoe UI",
-      fontStyle: "normal",
-      fontWeight: "normal"
     },
     matrixLabel: {
       paddingLeft: "20px"
@@ -97,37 +44,6 @@ export const matrixAreaStyles: () => IProcessedStyleSet<
     },
     matrixLabelTab: {
       width: "150px"
-    },
-    matrixRow: {
-      display: "flex",
-      flexDirection: "row",
-      fontFamily: "Segoe UI",
-      fontStyle: "normal",
-      fontWeight: "normal",
-      height: "50px"
-    },
-    nanMatrixCell: {
-      background:
-        "repeating-linear-gradient(-45deg, white, white 5px, pink 10px, pink 20px)",
-      color: "black"
-    },
-    selectedMatrixCell: {
-      alignItems: "center",
-      border: "3px solid #089acc",
-      display: "flex"
-    },
-    styledMatrixCell: {
-      ":focus": {
-        outline: "none"
-      },
-      alignItems: "center",
-      cursor: "pointer",
-      display: "flex",
-      fontSize: "18px",
-      height: "100%",
-      justifyContent: "center",
-      lineHeight: "24px",
-      width: "100%"
     }
   });
 };
